Fix script replace to handle single quotes and backslashes.
