+++ lib/fs-driver.js
@@ -30,7 +30,9 @@ class FsDriver {
     if (!isString(str)) {
       str = str.toString()
     }
-    return str.replace(/(['/\\])/g, '\\$1')
+    return str.replace(/'/g, '\'"\'"\'')
+      .replace(/[\\]/g, '\\\\')
+      .replace(/([/])/g, '\\$1')
   }
 
   /**
@@ -281,13 +283,6 @@ class FsDriver {
           self.root + '$1'
         )
       }
-
-      if (err && err.code === 'ENOENT') {
-        err.cmd = cmd
-        err.args = args
-        err.scriptCommand = scriptCommand
-      }
-
       cb(err, output, scriptCommand)
     })
   }
