+++ test/fixtures/root/E
@@ -1,0 +1,3 @@
+File E
+
+'username' => ''
