+++ test/functional/transformer.js
@@ -4,17 +4,11 @@ var Lab = require('lab')
 var lab = exports.lab = Lab.script()
 var describe = lab.describe
 var it = lab.it
-var before = lab.before
 var beforeEach = lab.beforeEach
-var after = lab.after
 var afterEach = lab.afterEach
 var Code = require('code')
 var expect = Code.expect
 var fs = require('../fixtures/fs-helper')
-var async = require('async')
-var childProcess = require('child_process')
-
-var debug = require('debug')('fs-transform:test')
 
 var Transformer = require('../../index')
 
@@ -190,6 +184,23 @@ describe('functional', () => {
         done()
       })
     })
+
+    it('should handle single quotes', (done) => {
+      var rules = [
+        {
+          action: 'replace',
+          search: '\'username\' => \'\'',
+          replace: '\'username\' => \'doobs\''
+        }
+      ]
+      Transformer.transform(fs.path, rules, (err, transformer) => {
+        expect(err).to.not.exist()
+        let lines = transformer.results[0].diffs['/E'].split('\n')
+        expect(lines[5]).to.equal('-\'username\' => \'\'')
+        expect(lines[6]).to.equal('+\'username\' => \'doobs\'')
+        done()
+      })
+    })
   }) // end 'replace'
 
   describe('results', () => {
@@ -260,106 +271,4 @@ describe('functional', () => {
       })
     })
   }) // end 'results'
-
-  describe('scripts', () => {
-    const scriptPath = fs.mock + '.script'
-
-    before(fs.createDotGit)
-    after(fs.removeDotGit)
-
-    beforeEach((done) => {
-      childProcess.exec('cp -r ' + fs.mock + ' ' + scriptPath, done)
-    })
-
-    afterEach((done) => {
-      var command = 'rm -rf ' + scriptPath
-      childProcess.exec(command, {cwd: 'test/fixtures/'}, done)
-    })
-
-    function compareScript (rules, done) {
-      var script
-      async.series([
-        function generateScript (next) {
-          Transformer.transform(fs.path, rules, (err, transformer) => {
-            if (err) { return next(err) }
-            script = transformer.getScript()
-            debug(script)
-            fs.writeFile(scriptPath + '/script.sh', script, next)
-          })
-        },
-        function runScript (next) {
-          childProcess.exec('bash script.sh', { cwd: scriptPath }, (err, output) => {
-            debug(output)
-            next(err)
-          })
-        },
-        function removeScript (next) {
-          childProcess.exec('rm -f ' + scriptPath + '/script.sh', next)
-        },
-        function getDiff (next) {
-          var command = 'diff -r ' + fs.path + ' ' + scriptPath
-          childProcess.exec(command, (err, diff) => {
-            if (err && err.code > 1) { return next(err) }
-            debug(diff)
-            expect(diff).to.be.empty()
-            next()
-          })
-        }
-      ], done)
-    }
-
-    it('should provide a shell script correctly transforms', (done) => {
-      var rules = [
-        { action: 'replace', search: '\\sum', replace: '\\prod' },
-        { action: 'copy', source: 'A', dest: 'A-copy' },
-        { action: 'copy', source: 'B', dest: 'B-copy' },
-        { action: 'rename', source: 'sub/C', dest: 'sub/C-rename' }
-      ]
-      async.series([
-        function runScript (next) {
-          var command = 'bash ../../fixtures/script.sh'
-          childProcess.exec(command, {cwd: scriptPath}, (err, data) => {
-            next(err)
-          })
-        },
-
-        function runTransforms (next) {
-          Transformer.transform(fs.path, rules, next)
-        },
-
-        function getDiff (next) {
-          var command = 'diff -r ' + fs.path + ' ' + scriptPath
-          childProcess.exec(command, (err, diff) => {
-            if (err && err.code > 1) { return next(err) }
-            expect(diff).to.be.empty()
-            next()
-          })
-        }
-      ], done)
-    })
-
-    it('should correctly handle global excludes', (done) => {
-      compareScript([
-        { action: 'exclude', files: ['sub/C', 'A'] },
-        { action: 'replace', search: 'Mew', replace: 'Woof' }
-      ], done)
-    })
-
-    it('should handle local replace excludes', (done) => {
-      compareScript([
-        {
-          action: 'replace',
-          search: 'Mew',
-          replace: 'Bark',
-          exclude: ['sub/C']
-        }
-      ], done)
-    })
-
-    it('should always exclude .git files', (done) => {
-      compareScript([
-        { action: 'replace', search: 'only_in_gitfile', replace: 'yus' }
-      ], done)
-    })
-  }) // end 'scripts'
 }) // end 'functional'
